# app/api/openapi.yaml
openapi: 3.0.0
info:
  title: AWS Services API
  version: 1.0.0
  description: Comprehensive API for managing AWS services including Lambda, API Gateway, IAM, SNS, SQS, and Step Functions
servers:
  - url: /api
    description: API v1
components:
  securitySchemes:
    AWS4-HMAC-SHA256:
      type: apiKey
      in: header
      name: Authorization
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string

    # S3 Schemas
    S3Bucket:
      type: object
      properties:
        Name:
          type: string
        CreationDate:
          type: string
        Region:
          type: string
        Versioning:
          type: boolean
        Encryption:
          type: string
          enum: [AES256, aws:kms]

    # Lambda Schemas
...
    SNSTopic:
      type: object
      properties:
        Name:
          type: string
        DisplayName:
          type: string
        Policy:
          type: object
        DeliveryPolicy:
          type: object
        Tags:
          type: array
          items:
            type: object
            properties:
              Key:
                type: string
              Value:
                type: string
...
paths:
  /s3:
    get:
      summary: List S3 buckets
      operationId: listBuckets
      responses:
        '200':
          description: Successfully retrieved buckets
          content:
            application/json:
              schema:
                type: object
                properties:
                  buckets:
                    type: array
                    items:
                      $ref: '#/components/schemas/S3Bucket'
                  requestId:
                    type: string
                  timestamp:
                    type: string
                  region:
                    type: string
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create S3 bucket
      operationId: createBucket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/S3Bucket'
      responses:
        '201':
          description: Bucket created successfully
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /sns/topics:
    get:
      summary: List SNS topics
      operationId: listTopics
      responses:
        '200':
          description: Successfully retrieved topics
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      $ref: '#/components/schemas/SNSTopic'
                  requestId:
                    type: string
                  timestamp:
                    type: string
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create SNS topic
      operationId: createTopic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SNSTopic'
      responses:
        '201':
          description: Topic created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SNSTopic'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: AWS Authentication Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...